+++ apps/nizam/app/[locale]/decks/[id]/cards/actions.ts
@@ -1,0 +1,72 @@
+'use server'
+
+import { revalidatePath } from 'next/cache'
+import { CreateFlashcardDtoTypeEnum, createServerTedrisatAPIs } from '@madrasah/services/tedrisat'
+import { env } from '~/env'
+import { auth } from '~/lib/auth_options'
+
+export const updateFlashcard = async (cardId: number, updatedCard: {
+  contentFront?: string
+  contentBack?: string
+}) => {
+  const session = await auth()
+  const { cards } = await createServerTedrisatAPIs(session?.accessToken, env.TEDRISAT_API_BASE_URL)
+
+  try {
+    await cards.updateFlashcard({
+      id: cardId,
+      updateFlashcardDto: {
+        contentBack: updatedCard.contentBack,
+        contentFront: updatedCard.contentFront,
+      },
+    })
+
+    return true
+  }
+  catch (error) {
+    console.log('Error updating flashcard:', error)
+    return false
+  }
+}
+
+export const createFlashcards = async (deckId: number, newCards: {
+  contentFront: string
+  contentBack: string
+}[]) => {
+  const session = await auth()
+  const { cards } = await createServerTedrisatAPIs(session?.accessToken, env.TEDRISAT_API_BASE_URL)
+
+  try {
+    const response = await cards.createFlashcards({
+      deckId,
+      createFlashcardDto: newCards.map(card => ({
+        type: CreateFlashcardDtoTypeEnum.Hadeeth,
+        contentFront: card.contentFront,
+        contentBack: card.contentBack,
+      })),
+    })
+    revalidatePath(`/decks/${deckId}/cards`)
+    return response
+  }
+  catch (error) {
+    console.log('Error creating flashcards:', error)
+    return false
+  }
+}
+
+export const deleteFlashcard = async (cardId: number, deckId?: number) => {
+  const session = await auth()
+  const { cards } = await createServerTedrisatAPIs(session?.accessToken, env.TEDRISAT_API_BASE_URL)
+
+  try {
+    await cards.deleteFlashcard({ id: cardId })
+
+    // Revalidate the cards page to refresh server-side data
+    revalidatePath(`/decks/${deckId}/cards`)
+    return true
+  }
+  catch (error) {
+    console.log('Error deleting flashcard:', error)
+    return false
+  }
+}