+++ apps/tedris/features/flashcards/components/flashcard-list.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { CaretLeftIcon, CaretRightIcon } from '@madrasah/icons'
-import { useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 
 import FlashCardContent from './flashcard-content'
 import { FlashcardResponse } from '@madrasah/services/tedrisat'
@@ -14,29 +14,47 @@ export default function FlashCardList({ cards }: FlashCardListProps) {
   const [currentIndex, setCurrentIndex] = useState(0)
   const [key, setKey] = useState(0)
 
-  const handlePrevious = () => {
+  const handlePrevious = useCallback(() => {
     setCurrentIndex(prev => (prev > 0 ? prev - 1 : cards.length - 1))
     setKey(prev => prev + 1)
-  }
+  }, [cards.length])
 
-  const handleNext = () => {
+  const handleNext = useCallback(() => {
     setCurrentIndex(prev => (prev < cards.length - 1 ? prev + 1 : 0))
     setKey(prev => prev + 1)
-  }
+  }, [cards.length])
+
+  useEffect(() => {
+    const onKeyDown = (e: KeyboardEvent) => {
+      if (e.key === 'ArrowLeft') {
+        handlePrevious()
+      }
+      else if (e.key === 'ArrowRight') {
+        handleNext()
+      }
+    }
+
+    window.addEventListener('keydown', onKeyDown)
+
+    return () => {
+      window.removeEventListener('keydown', onKeyDown)
+    }
+  }, [cards.length, handleNext, handlePrevious])
 
   if (!cards.length) {
     return (
-      <div className="flex h-[500px] items-center justify-center">
+      <div className="flex items-center justify-center">
         <p className="text-gray-500">Henüz kart bulunmuyor.</p>
       </div>
     )
   }
 
   return (
-    <div className="mx-auto flex w-full max-w-3xl flex-col items-center justify-center">
+    <div className="mx-auto relative h-full w-full max-w-3xl">
       {cards[currentIndex] && (
         <FlashCardContent key={key} {...cards[currentIndex]} />
       )}
+
       <div className="mt-4 flex items-center justify-between">
         <button
           onClick={handlePrevious}