+++ apps/tedris/next.config.js
@@ -1,6 +1,12 @@
+import createNextIntlPlugin from 'next-intl/plugin'
+const withNextIntl = createNextIntlPlugin('./lib/i18n/request.ts')
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   transpilePackages: ['@madrasah/ui', '@madrasah/utils', '@madrasah/types', '@madrasah/hooks', '@madrasah/services'],
+  experimental: {
+    optimizePackageImports: ['@madrasah/icons', '@madrasah/icons/ssr', '@phosphor-icons/react'],
+  },
   redirects: async () => {
     return [
       {
@@ -20,4 +26,4 @@ const nextConfig = {
   },
 }
 
-export default nextConfig
+export default withNextIntl(nextConfig)