+++ apps/nizam/CHANGELOG.md
@@ -1,5 +1,45 @@
 # Changelog
 
+## [0.1.7](https://github.com/amel-tech/madrasah-frontend/compare/nizam-web-v0.1.6...nizam-web-v0.1.7) (2025-12-06)
+
+
+### Features
+
+* added uthman font ([091232b](https://github.com/amel-tech/madrasah-frontend/commit/091232b88ac0ebf2a4ba9d7996a50aa573acd516))
+* added uthman font ([fb7488e](https://github.com/amel-tech/madrasah-frontend/commit/fb7488e3291ca25b9b74260f77523a7a9d71d36f))
+* **authentication:** add refresh token rotation ([efd21b6](https://github.com/amel-tech/madrasah-frontend/commit/efd21b6f53860bbb038e9fe5f4cf486d600e950d))
+* **authentication:** add route protection and redirect if not authenticated ([5d08bea](https://github.com/amel-tech/madrasah-frontend/commit/5d08bea98f5edb588db7babbdc9358e8e68bf38a))
+* **authentication:** simplify, session obtaining by a helper func ([71bc996](https://github.com/amel-tech/madrasah-frontend/commit/71bc996ad32c3e639599ab117f027d6d4d993d26))
+* creating i18n package ([8664c1d](https://github.com/amel-tech/madrasah-frontend/commit/8664c1dba4bc9b691304696891a9c5eba1850d68))
+* **i18n:** integrate NextIntlClientProvider into layout components ([17df792](https://github.com/amel-tech/madrasah-frontend/commit/17df7924e15a8803f7b5c9498791a261cac34f81))
+* implementing next-intl package on nizam ([ebe3922](https://github.com/amel-tech/madrasah-frontend/commit/ebe392294eb9b51de3f07cf410d1630ff546347d))
+* **nizam:** adding i18n local switcher ([cf8f72c](https://github.com/amel-tech/madrasah-frontend/commit/cf8f72cfa0efff66011fc716fcaae572c3f18e1c))
+
+
+### Bug Fixes
+
+* **authentication:** use session.accesToken instead of session-token, which was jwe token ([5d7d556](https://github.com/amel-tech/madrasah-frontend/commit/5d7d556b601793f51b5cc12e2c9d6e9aad0cbd74))
+* **authentication:** use session.accesToken instead of session-token,… ([7d40a3f](https://github.com/amel-tech/madrasah-frontend/commit/7d40a3f1d55accd1e30dc963c26add318faaf58d))
+* update local urls and expose new ports for nazar, nizam, and ted… ([5f6b16f](https://github.com/amel-tech/madrasah-frontend/commit/5f6b16f414151d72455e2be935cf64ec811e449a))
+* update local urls and expose new ports for nazar, nizam, and tedris applications ([cd87b5f](https://github.com/amel-tech/madrasah-frontend/commit/cd87b5ff5e1a1491a1486324e20b2a8af89a0869))
+
+## [0.1.6](https://github.com/amel-tech/madrasah-frontend/compare/nizam-web-v0.1.5...nizam-web-v0.1.6) (2025-10-30)
+
+
+### Features
+
+* **flashcards:** change tedris flashcards edit screen to table ([ca15f63](https://github.com/amel-tech/madrasah-frontend/commit/ca15f6376dd4b8daf9c2532a2e24c7f4bf1a6dff))
+* **flashcard:** turn create-deck action into dialog ([f383604](https://github.com/amel-tech/madrasah-frontend/commit/f383604588952240bb59870eadb5c0cb5ea34dbd))
+* **nizam:** using the new editable columns ([0562fbc](https://github.com/amel-tech/madrasah-frontend/commit/0562fbc73df46812061d67b496de92327d916299))
+
+## [0.1.5](https://github.com/amel-tech/madrasah-frontend/compare/nizam-web-v0.1.4...nizam-web-v0.1.5) (2025-10-21)
+
+
+### Bug Fixes
+
+* revert-nizam-port-change ([ebc6af5](https://github.com/amel-tech/madrasah-frontend/commit/ebc6af5b396f59235396efbeb52cd825d6a2631a))
+* revert-nizam-port-change ([0fb493d](https://github.com/amel-tech/madrasah-frontend/commit/0fb493de6d7203939b11ab5a69983a97a9021deb))
+
 ## [0.1.4](https://github.com/amel-tech/madrasah-frontend/compare/nizam-web-v0.1.3...nizam-web-v0.1.4) (2025-10-16)